Dijkstra: Shortest Reach 2
Given a graph consisting  nodes (labelled  to ) where a specific given node  represents the starting position  and an
edge between two nodes is of a given length, which may or may not be equal to other lengths in the graph.

It is required to calculate the shortest distance from the start position (Node ) to all of the other nodes in the
graph.

Note: If a node is unreachable , the distance is assumed as .